remote test execution: decouple result.setup() from copying of tests
The result.setup() method is now only used to setup the communication
channel, and the copy of the tests will be made conditional, based on
the presence of Avocado on the remote machine.

On the unittest side, I'm quite adverse to heavy mocking, so I favored
removing all these painful, and IMHO of little value checks.

Signed-off-by: Cleber Rosa <crosa@redhat.com>