avocado.core.loader: Remove job from test loader
Test loader doesn't need to know anything about the job, the only useful
argument is `args`.

Second big change is the way loader is used. This patch uses list of
registered plugins on module-level (similarly to settings or logging)
instead of custom instance and plugins registered in args dict. This way
the plugins are loaded in the order they are registered (by plugin
priority) instead of random order.

Signed-off-by: Lukáš Doktor <ldoktor@redhat.com>