avocado.core.test: Test() -> Enforce that skip() can only be called from setUp()
Make sure avocado complains loudly when people try to call
skip() from inside setUp() - this shouldn't be allowed in
avocado tests. Let the user know of the problem and ask them
to fix their tests. Also, update the docstring for skip(),
which will reduce the chances of test writers using it in
inappropriate places.

Signed-off-by: Lucas Meneghel Rodrigues <lmr@redhat.com>