avocado.core.job: Start logging earlier and add root_logger by default
Avocado defines own logging streams, all non-matching streams should
go to the job log.

This patch enables logging right after we setup the result direcotry to
allow the job to log early messages for example from test-discovery
directly into the job log.

Additionally this patch increases the module column in logging as lots
of module names are longer than 8 chars.

As the last change this removes the "avocado.linux" stream as it not
used anywhere.

Signed-off-by: Lukáš Doktor <ldoktor@redhat.com>