GDB Plugin: Change undefined behavior exception
When we changed the policy regarding skipping tests,
we forgot to change how the gdb plugin behaves on
undefined behavior - We are not supposed to be skipping
tests during its execution anymore.

So let's fix the problem by changing the undefined
behavior exception to a TestError instead of a TestNAError.

Signed-off-by: Lucas Meneghel Rodrigues <lmr@redhat.com>