avocado.utils.astring: Remove trailing whitespaces from tabular output
We did notice that listing commands have trailing whitespace
at their ends. Trailing whitespaces are bad for program output
(copy/paste, among other reasons), so let's fix it.

I did notice there was no reason to not implement the base
function for the tabular output as a generator, so I've
introduced iter_tabular_output, that is a generator, and
in order to not break API unnecessarily, implemented
tabular_output in terms of iter_tabular_output.

Signed-off-by: Lucas Meneghel Rodrigues <lmr@redhat.com>