avocado.core.multiplexer: Assign tag to match multiplex variant
"avocado multiplex" and "avocado run" should share the same variant
numbers. This patch forces multiplexer to set the tag when generating
multiple variants. This way the tag is always set accordingly to the
variant.

When multiple tests of the same name are executed, the first argument is
the multiplexed variant number and additional `.$num` is added to
generate unique name.

Signed-off-by: Lukáš Doktor <ldoktor@redhat.com>