avocado.core.runner: Make getting the timeout from test plugins more robust
Some test loader plugins will return the 'timeout' parameter
as a string, so this will cause avocado to break, given that
it will add the timeout to another float value later in the
runner code. Let's ensure that value will be converted to
a float before the addition is made.

Signed-off-by: Lucas Meneghel Rodrigues <lmr@redhat.com>