VM Plugin: no need to gracefully shutdown VM if it's state will be reverted
Let's be practical and simply destroy (pull the plug on the VM) that
will have its state (snapshot reverted).

Not shutting down (or powering off/destroying) the VM has the side
effect of not reverting the state of the (still running) VM on a
couple of platforms, including EL7.

This fixed issue #396.

Signed-off-by: Cleber Rosa <crosa@redhat.com>