avocado/utils/kernel_build.py: make config optional
A Linux kernel can be built without an external configuration file
because it can generate a default one for the current target
platform. So, let's make that parameter optional.

Also, if a user does provide a config file, we use a completely silent
target for updating that config file, using defaults for missing
configuration keys.

Signed-off-by: Cleber Rosa <crosa@redhat.com>