avocado.core.runner: Change the throbber colors
Throbber (test progress) supports two colors. Currently it turns green
only when new status arrives, which is currently only used when we
interrupt the test (GDB) or when the test finishes. Then the throbber
turns green for one iteration and back to yellow in next iteration.

This patch changes the behavior so the throbber is green when the test
process is alive and yellow when it's interrupted.

Signed-off-by: Lukáš Doktor <ldoktor@redhat.com>