avocado.core.output.View: Display error messages on paginator
The fix in c514013e924f5a2c8523edc20418adb9d4638f30 was
only partially successful, because I forgot that
the paginator should also get important error conditions
displayed by the program. So let's fix that by changing
the display logic (when an error message is displayed,
do it in the paginator and in the app logger).

Signed-off-by: Lucas Meneghel Rodrigues <lmr@redhat.com>