Selftests: allow tests to be run with "setup.py test"
One standard setuptools command is "test". That command does some useful
setup and then runs the test in a verbose mode.

There are a couple of reasons for activating that, including that
setuptools can help us support a Python 3 version of Avocado by
running 2to3 and even custom fixers before running the tests, or even
builds and installs.

Signed-off-by: Cleber Rosa <crosa@redhat.com>