Selftests: drop nose usage for plain unittest discovery
The unittest module, and unittest2 backport on Python 2.6, is capable of
finding unittests and running them. So, effectively, we do not need nose
at all.

This patch removes the dependency on nose, replacing the run script
with a version based solely on the unittest module.

One change of functionality is that run now looks and runs tests in the
standard selftests directories (unit, functional, doc), and does not
accept command line arguments. If one wants to run a subset of them,
it's pretty easy to just use the unittest module for that:

 $ python -m unittest discover -s selftests/unit

Signed-off-by: Cleber Rosa <crosa@redhat.com>