setup.py: Set zip_safe flag to false
In module avocado.core.plugins.builtin, There is a calling of
os.listdir() to iterate files in the installed python directory.

If avocado is installed in an .egg file which is compressed, it will
fail to load because that path does not really exists.

To avoid this failure, this patch set zip_safe to False explicitly.
So setuptools will install package to a real directory instead of a
zipped .egg file.

Signed-off-by: Hao Liu <hliu@redhat.com>