Plugins refactor: define interfaces and rely on stevedore for loading
This introduces a complete (WiP) refactor of the plugin architecture.
There are two big aspects here:

1) Plugin interfaces are now documented and live under 'avocado.plugins'.
2) The mechanism for loading/activating/running plugins is based on stevedore,
   a flexible and stable Python library used in many other projects, namely
   OpenStack projects.

For now, as a WiP, this adds this new mechanism as a secondary plugin loading
mechanism. A couple of "dispatcher" classes were created alongside the current
plugin manager.

Of the differences is that each dispatcher represents one specific Plugin
interface definition, such as the class "avocado.plugins.base.CLIRUNBase",
which allows one to add arguments to the command line interface "run" command,
or "avocado.plugins.base.ResultBase", which allows plugins to output to
different formats.

Once the migration is completed, the current plugin manager can be replaced
for the "dispatchers" introduced here.

Signed-off-by: Cleber Rosa <crosa@redhat.com>