avocado.core.plugins: Print traceback on failure
On plugin import failure we get the error message, which in case of
IOError is: "IOError: [Errno 2] No such file or directory". We have to
debug it or add some prints to discover, what happened. Let's add
optional traceback print in the same fashion as the "loader" does.
(the output is ignored unless the user enables "avocado.app.tracebacks")

Signed-off-by: Lukáš Doktor <ldoktor@redhat.com>