avocado.spec: Temporarily disable check due to problems on COPR builds
After a lot of work, I've reduced the amount of build problems
due to miscellaneous problems in different architectures. I couldn't
get the amount of failures to 0 (I have a pretty detailed report
to continue work on this though), so in order to have a decent
0.28.0 build I'll disable the check stage for now. I'll revert
the change once I get all chroots to build cleanly.

Signed-off-by: Lucas Meneghel Rodrigues <lmr@redhat.com>