selftests/test_basic.py: Put tmp dir checks at RunnerOperationTest
Add a verifyAvocadoTempDirs method to RunnerOperationTest so
that we can verify that there are no /var/tmp/avocado_* dirs
left at the end of tests. Also, add one test where we execute
avocado --help and verify if there's such a dir at the end
of the test. This will make us less likely to regress the
leftover /var/tmp/avocado_* dirs after --help is called.

Signed-off-by: Lucas Meneghel Rodrigues <lmr@redhat.com>