avocado.core.remote.runner: Fix incorrect test link in the HTML report
It turns out we were basing the links of the URL report
in the 'url' attribute, when we should be using 'test'.
This fixes an issue that came up with the new multiplexed
tests in VM and remote plugins, where the HTML report
link was invalid. Fix the unittest, that had the wrong
behavior built into it.

Signed-off-by: Lucas Meneghel Rodrigues <lmr@redhat.com>