avocado.core.data_dir: Ensure tmpdir is there before accessing it
In order to avoid re-creating the temporary directory,
put a sanity check on the .get_tmp_dir() function. This
avoids re-creating temporary directories, which is a
security breach. Put the appropriate handling in the
clean tmp files function, given that when we want to
clean up the directory, if the dir is not there anymore
by definition there's nothing to be cleaned up.

Signed-off-by: Lucas Meneghel Rodrigues <lmr@redhat.com>