avocado: Turn test.Test.workdir into a LazyProperty
Turn .workdir and .srcdir properties of the test.Test
object into lazily instantiated properties (that means
they'll only be created when they are actually accessed).

This means we won't create extraneous directories in our
functional and unit tests for the avocado test classes
themselves. In this commit, we introduce
data_structures.LazyProperty, then use it in test.Test
to define the aforementioned properties.

Signed-off-by: Lucas Meneghel Rodrigues <lmr@redhat.com>