VM Plugin: make --vm-hostname optional by attempting to find IP address
By leveraging the `avocado.core.virt.VM.ip_address` method, let's make
the `--vm-hostname` optional. If the IP address can not be found, then
ask users to manually provide it.

If the user has a VM running, or has had it running a "while" back,
chances are pretty high Avocado will be able to find it's IP address.

Signed-off-by: Cleber Rosa <crosa@redhat.com>