avocado/core/remote/result.py: replace TestSetupFail for JobError
According to the basic understanding of the wording "TestSetupFail"
doesn't here apply here. The misuse of the exception is further
confirmed by the docstring: "Indicates an error during a setup or
cleanup procedure."

Not being able to connect/start/snapshot the VM to be used to run all
tests are not related to the test setup/cleanup steps. JobError seems
to be a more appropriate exception to be used here.

Signed-off-by: Cleber Rosa <crosa@redhat.com>