Test Loader: test class is a string for avocado instrumented tests
With the changes that now do not load the test at discovery time,
the test class is a string with the class name, and not a class
itself.

Therefore, this discovery test is only going to check for the test
class name. Other test may be added to do the real module load and
test class instantiation.

Signed-off-by: Cleber Rosa <crosa@redhat.com>