Test Loader: do not load/execute test modules at discovery time
Currently, the test loader will load/execute the Python test files at
discovery time. This is not desirable because, among other things,
it's unsafe to execute code that was not explicitly requested to be
executed.

The new approach is to use the `ast` module and parse, but not
execute, the Python source code files, and attempt to use some
heuristics to detect an Avocado INSTRUMENTED test.

Since the current loader is too tighthly coupled to the idea of having
test classes and parameters (which implies having loaded the code) as
the building blocks of tests, this new approach gives the system class
names only. Also, since the code is not loaded, the path to the source
code file is given inside the test parameters (and later pulled
out). A more complete solution would be to redesign the Loader so that
it gives test information only at discovery time for *all* test types.

Since there's another major work going on in the loader at this time,
I decided to keep things simple(r) and just deal with the string based
"classes" when necessary.

Also included (transparently) is the support for multiple classes with
multiple tests inside a single Python source code file.

Signed-off-by: Cleber Rosa <crosa@redhat.com>