avocado.core.data_dir: Make logdir creation safer
When one execute 2 jobs with first 7 chars of unique id at the same
time, they end-up in the same logdir.

This patch makes the creation safer. First it creates the main
directory, then it tries to create the directory with 7 chars of the
unique id. On failure it adds another char until it finds empty dir.

In case you spawn 2 jobs with the same id at the same time, it adds
a number starting with 0.

If you spawn more tham 1000 jobs at the same time with the exact same
id, it fails with IOError.

Signed-off-by: Lukáš Doktor <ldoktor@redhat.com>