avocado.core.tree: Improve error message on missing !include file
When a multiplex file includes a missing file, avocado only reports
"there is a missing file" without any explanations whatsoever. This
patch changes it to a ValueError with the original file and the
missing one.

Additionally it adds missing Exception to try/except, which was
already possible on certain errors, to avoid tracebacks.

Signed-off-by: Lukáš Doktor <ldoktor@redhat.com>