docs: remove static modules.png file and its reference in the docs
The modules.png file is a static file that would need manual update
with almost every single commit. This is rather impractical and
doesn't add that much value to the regular user.

If a developer wants to take a look at the dependency graph, a
Makefile target still exists that will give a user a fresh graph.

Signed-off-by: Cleber Rosa <crosa@redhat.com>