avocado.core.loader: Use only correct test_types
The usage usage of incorrect test-types (types which can't be executed
like missing file, broken symlink ...) was not found useful. This
patch only allows "good" test types to be listed and possible to use.

Additionally the test type now requires to specify loader and type to
avoid possible clashes (and to emphasize the relation between loader
and test type).

Signed-off-by: Lukáš Doktor <ldoktor@redhat.com>