avocado.core.loader: Refactor inner_runner to be a separate loader
Currently InnerRunner lives within the FileLoader plugin and overrides
it's function in case inner_runner is supplied. This patch extracts
it and creates separate plugin for it. This plugin is still part of the
core and is registered automatically during the runtime.

On --inner-runner usage it overrides the --loader order and registers
only itself as by specifying --inner-runner people asks to use it.

When inner_runner is specified in --loader, it's processed accordingly
to the priority. When one uses only "inner_runner" without extra_params,
it never matches any tests (to keep the behavior).

Signed-off-by: Lukáš Doktor <ldoktor@redhat.com>