Doc test: check condition for building the docs with Sphinx
Sphinx will attempt to fetch the Python objects inventory during the build
process. If for some reason, this test is being run on a machine that can
not access that address simply because of network restrictions (or the
developer may simply be on a plane) then it's better to SKIP the test than
to give a false positive.

Changes from v0:
 * Use unittest2 on Python 2.6

Signed-off-by: Cleber Rosa <crosa@redhat.com>