Makefile: Use mock to build RPMs
Instead of trusting in rpmbuild to do the rpm builds,
use mock - a program that builds the rpms from fresh
chroots. This ensures that BuildRequires are correct
and that the binary rpms build cleanly.

This requires a few changes to the Makefile targets
and the avocado spec file, laid out in this patch.

Signed-off-by: Lucas Meneghel Rodrigues <lmr@redhat.com>