avocado: Run plain python unittests as simple tests
Make python unittests to run as simple tests even if
they are not executable. Use a very simple, ast based
inspection to see if the python file is a unittest.
Inside simple test, check if the file is a non executable
python file, if so, append the python interpreter to
its command.

Signed-off-by: Lucas Meneghel Rodrigues <lmr@redhat.com>