avocado.utils.software_manager: Add DNF as a backend
From Fedora 22 onwards, DNF is the official package
manager of that distro. Although mostly similar to
yum, there are a few differences that must be taken
to account (besides, the yum wrapper that ships with
dnf is going to go deprecated eventually).

So introduce a basic DnfBackend class and detection
to software manager.

Signed-off-by: Lucas Meneghel Rodrigues <lookkas@gmail.com>