RFC: location discovery
Right now, Avocado does not have a clear separation during the test
discovery process of what can be a valid test location (think of a
test repository or collection) and, in a more detailed fashion, the
tests (and their types) that could be found in those locations.

One trivial but valid example: a single Python source code file can be
a valid *location* for many individual tests. These tests may be
scattered around many different classes. These classes, may actually
implement tests that Avocado may recognize (and treat) differently,
such as Avocado Instrumented tests or standard Python unittest
TestCase based tests.

This may sound complex, simply because it actually is. IMHO, the
solution to complex is to replace it with something simpler.

This RFC proposes a simple idea to cover the first stage: discoverying
*possible* test locations. Once that is clear and separated enough in
Avocado code, then it should be simpler to support the next stages.

The discovery code intentionally shows different classes that
"discover" test locations based on any files (thing of discovering
/usr/bin/) and then later a class that only cares about Python source
code files. The intention is to be flexible enough to allow for any
custom location discovery, such as:

 * It must be a file must be executable (a valid case for simple tests)
 * It must be reachable over HTTP and its name must end with ".js"

A test "inspector", or something better named, should then receive a list
of locations and detect if they actually contain valid tests, and their
specific types.

The idea is to have a reasonable (but configurable) list of location discovery
classes. Also, with the upcoming plugin refactor, users should be able to enable,
disable and configure individual location discovery classes themselves.

Signed-off-by: Cleber Rosa <crosa@redhat.com>