avocado.core.test: Avoid problems with '/' in log-dir
We base our log-dir on test name, which can contain various characters
including '/', causing multiple directories being created instead of one
log directory. This patch modifies the `self.logdir` to use '_' instead
'/'.

Note this approach is not safe on windows so in the future we might
implement a `utils` function to make filenames "safe". (similarly to
django, which uses whitelist)

Signed-off-by: Lukáš Doktor <ldoktor@redhat.com>