avocado.utils.software_manager: Add local file install support to AptBackend
It turns out that apt doesn't really handle local installs
graciously as dnf/yum, but thankfully there is a program
named gdebi, shipped in Debian and Ubuntu, that would let
us do that. Use it when available (we can make gdebi a
dependency for the avocado debian package to ensure it'll
always be there).

Signed-off-by: Lucas Meneghel Rodrigues <lookkas@gmail.com>