avocado/core/virt.py: avoid crashing when trying to detect VM IP address
The current code assumes that libvirt will always return information
about the VM in a given format, and also assumes that the VM will have
a NIC, which may not be the case.

This adds better checks if the conditions are not met, skipping the
IP detection feature, instead of crashing.

Signed-off-by: Cleber Rosa <crosa@redhat.com>