process.py: CmdError class backward compatibility fix
In the newer CmdError class implementation the __init__() doesn't
have anymore the third non-mandatory argument "additional_text".
But still in the different parts of the source code the
CmdError is used with the third argument. If there is no special
reason behind deleting the "additional_text" then I would suggest
to keep for having the backward compatibility.
(for example got argument amount errors while running particular
libvirt tests which had to raise a CmdError exception)