GDB support in process: be more robust against malformed commands
It has been reported that avocado crashes when a malformed command
is attempted to be executed, even when using `shell=True`.

The reason is that `shlex.split` will raise `ValueError` on those
commands. Since improving `shlex.split` may be too complex, and
possibly not the best place to "fix" this, let's just assume safely
that this command should be run on GDB.

Signed-off-by: Cleber Rosa <crosa@redhat.com>