avocado.core.remoter: Fix error on exception reference
fabric.network was never imported in avocado.core.remoter,
but fabric.network.NetworkError was referenced in its code,
making the exception to never actually be captured. Let's
fix it.

While at it, let's resolve one complaint of the pycharm
python checker: When importing some module on a try/except
block, make some effort to define the module in case of
failure.

Signed-off-by: Lucas Meneghel Rodrigues <lookkas@gmail.com>