avocado.utils.service: Make ServiceManager available to non-root
On local runs, for non-root users, it's possible to consult
the contents of /proc/1/cmdline. Of course, we should do
this only if we are on a local run (process.run is passed
to get_name_of_init) and we couldn't directly read from
/proc/1/exe.

After this patch, it's possible to instantiate ServiceManager
on local runs made by non-root users.

Signed-off-by: Lucas Meneghel Rodrigues <lookkas@gmail.com>