avocado.core.runner: Get proper status
Currently the test status was only updated on `TestStatus.check` method,
which was also used in `TestStatus.abort`. But the meaning of `check`
return is whether the job was interrupted. It does not mean the status
is valid, which in some cases might have reported the test status `None`.

This patch reworks the way we interact with `TestStatus`. It contains
`interrupt` variable, which means the test was interrupted and we should
leave the execution. It still have the `status` variable, which contains
the status, but now those two variables are automatically updated on
query (using __getattribute__).

Signed-off-by: Lukáš Doktor <ldoktor@redhat.com>