Makefile: make install-requirements[-selftests] best effort
The install requirements targets are really best effort. PIP can fail
in so many ways. So, let's not attempt to detect if the target ran
successfully or not. This is important because it may prevent other
chained command lines and targets to attempted an execution.

Also attempt to install Python 2.6 requirements early, since they are
simpler than the other requirements and pretty much basic back ported
libraries.

Signed-off-by: Cleber Rosa <crosa@redhat.com>