[TEMPORARY] command line parser: only show the basic usage info by default
This shows a more succint version of the help message (the so called
"usage" information). It's a certainly a matter of taste and change
of behavior, so this is marked as an RFC.

It also avoids a hackish injection of the print_help method as the
default dispatch method.

NOTE: This commit is going to be sent on a separate PR.

Signed-off-by: Cleber Rosa <crosa@redhat.com>