avocado.core: Initial job replay support
The initial job replay support using the run option "--replay <job-id>"
implements the job replay, retrieving the original job urls and multiplex.
The configuration from the original job, already recorded, is not implemented
at this point since it's still under discussion.

The option "--replay-test-status <FAIL,PASS>" is fully functional,
allowing users to replay only the variants that faced a given status in
the original job.

To ignore some original job information, the option
"--replay-job-ignore <mux|urls|config>" can be used. Notice if you inform the
urls, the multiplex or the configuration using the command line it will have
the same effect as in "--replay-job-ignore", making the replay job use the
provided information instead of the recorded information from original job.

Signed-off-by: Amador Pahim <apahim@redhat.com>