HTML output support: behave when the plugin is not installed
We currently only set HTML support as disabled if the `html.py`
library fails to import. The truth is that missing files that
are intended to be present on the (separately packaged) HTML
plugin package can cause other kinds of exceptions.

Let's add an extra check, looking if the resource files are
present (currently the template file).

Note: there's one outstanding bug, for which a fix is being
worked on. When Avocado is installed, either from packages
or from source code, the HTML plugin entry point is always
registered. That means the plugin Python module is always
tried to be loaded. If the plugin *package* is not installed,
an error message is given on the command line.

Signed-off-by: Cleber Rosa <crosa@redhat.com>